+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/ExcelhammerBlockDestroyedWithToolProcedure.java
@@ -34,9 +34,9 @@ public class ExcelhammerBlockDestroyedWithToolProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		if ((world.getBlockState(new BlockPos((int) x, (int) y, (int) z))).getMaterial() == net.minecraft.block.material.Material.ROCK
-				|| (world.getBlockState(new BlockPos((int) x, (int) y, (int) z))).getMaterial() == net.minecraft.block.material.Material.GLASS) {
-			world.destroyBlock(new BlockPos((int) x, (int) y, (int) z), false);
+		if ((world.getBlockState(new BlockPos(x, y, z))).getMaterial() == net.minecraft.block.material.Material.ROCK
+				|| (world.getBlockState(new BlockPos(x, y, z))).getMaterial() == net.minecraft.block.material.Material.GLASS) {
+			world.destroyBlock(new BlockPos(x, y, z), false);
 		}
 	}
 }
