+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/LilyPadValidPlacementConditionProcedure.java
@@ -34,10 +34,8 @@ public class LilyPadValidPlacementConditionProcedure {
 		double x = dependencies.get("x") instanceof Integer ? (int) dependencies.get("x") : (double) dependencies.get("x");
 		double y = dependencies.get("y") instanceof Integer ? (int) dependencies.get("y") : (double) dependencies.get("y");
 		double z = dependencies.get("z") instanceof Integer ? (int) dependencies.get("z") : (double) dependencies.get("z");
-		return (world.getBlockState(new BlockPos((int) x, (int) (y - 1), (int) z))).getMaterial() == net.minecraft.block.material.Material.WATER
-				&& !((world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z)))
-						.getMaterial() == net.minecraft.block.material.Material.WATER)
-				&& !((world.getBlockState(new BlockPos((int) x, (int) (y + 1), (int) z)))
-						.getMaterial() == net.minecraft.block.material.Material.LAVA);
+		return (world.getBlockState(new BlockPos(x, y - 1, z))).getMaterial() == net.minecraft.block.material.Material.WATER
+				&& !((world.getBlockState(new BlockPos(x, y + 1, z))).getMaterial() == net.minecraft.block.material.Material.WATER)
+				&& !((world.getBlockState(new BlockPos(x, y + 1, z))).getMaterial() == net.minecraft.block.material.Material.LAVA);
 	}
 }
