+++ src/main/java/io/github/team_lodestar/transcendeum/procedures/PerformExcerockAttackProcedure.java
@@ -123,7 +123,7 @@ public class PerformExcerockAttackProcedure {
 				}
 			}
 			if (world instanceof World && !world.isRemote()) {
-				((World) world).playSound(null, new BlockPos((int) x, (int) y, (int) z),
+				((World) world).playSound(null, new BlockPos(x, y, z),
 						(net.minecraft.util.SoundEvent) ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("entity.wither.break_block")),
 						SoundCategory.NEUTRAL, (float) 1, (float) 1);
 			} else {
